test(live): empty login回帰テストを既存suiteへ統合
Refs #1130 #1169
